requirements.txt
streamlit
pandas
pandas-profiling
streamlit-pandas-profiling
matplotlib
openai
pycaret
scikit-learn
python-dotenv
